feat: add connect wallet on cryptocurrency section, refactor styles and adjust tests
